+++ useScroll/useEvent.js
@@ -599,11 +599,15 @@ const useEventWheel = (props, instances) => {
 			positiveDelta = eventDelta > 0,
 			negativeDelta = eventDelta < 0,
 			{scrollTop, scrollLeft} = scrollContainerHandle.current;
-		let delta = 0;
+		let
+			delta = 0,
+			needToHideScrollbarTrack = false;
 
 		if (typeof window !== 'undefined') {
 			window.document.activeElement.blur();
 		}
+
+		scrollContainerHandle.current.showScrollbarTrack(bounds);
 
 		// FIXME This routine is a temporary support for horizontal wheel scroll.
 		// FIXME If web engine supports horizontal wheel, this routine should be refined or removed.
@@ -616,6 +620,7 @@ const useEventWheel = (props, instances) => {
 				// If ev.target is a descendant of scrollContent, the event will be handled on scroll event handler.
 				if (!utilDOM.containsDangerously(scrollContentRef.current, ev.target)) {
 					delta = scrollContainerHandle.current.calculateDistanceByWheel(eventDeltaMode, eventDelta, bounds.clientHeight * scrollWheelPageMultiplierForMaxPixel);
+					needToHideScrollbarTrack = !delta;
 
 					ev.preventDefault();
 				} else if (overscrollEffectRequired) {
@@ -623,8 +628,12 @@ const useEventWheel = (props, instances) => {
 				}
 
 				ev.stopPropagation();
-			} else if (overscrollEffectRequired && (negativeDelta && scrollTop <= 0 || positiveDelta && scrollTop >= bounds.maxTop)) {
-				scrollContainerHandle.current.applyOverscrollEffect('vertical', positiveDelta ? 'after' : 'before', overscrollTypeOnce, 1);
+			} else {
+				if (overscrollEffectRequired && (negativeDelta && scrollTop <= 0 || positiveDelta && scrollTop >= bounds.maxTop)) {
+					scrollContainerHandle.current.applyOverscrollEffect('vertical', positiveDelta ? 'after' : 'before', overscrollTypeOnce, 1);
+				}
+
+				needToHideScrollbarTrack = true;
 			}
 		} else if (canScrollHorizontally) { // this routine handles wheel events on any children for horizontal scroll.
 			if (negativeDelta && scrollLeft > 0 || positiveDelta && scrollLeft < bounds.maxLeft) {
@@ -633,11 +642,16 @@ const useEventWheel = (props, instances) => {
 				}
 
 				delta = scrollContainerHandle.current.calculateDistanceByWheel(eventDeltaMode, eventDelta, bounds.clientWidth * scrollWheelPageMultiplierForMaxPixel);
+				needToHideScrollbarTrack = !delta;
 
 				ev.preventDefault();
 				ev.stopPropagation();
-			} else if (overscrollEffectRequired && (negativeDelta && scrollLeft <= 0 || positiveDelta && scrollLeft >= bounds.maxLeft)) {
-				scrollContainerHandle.current.applyOverscrollEffect('horizontal', positiveDelta ? 'after' : 'before', overscrollTypeOnce, 1);
+			} else {
+				if (overscrollEffectRequired && (negativeDelta && scrollLeft <= 0 || positiveDelta && scrollLeft >= bounds.maxLeft)) {
+					scrollContainerHandle.current.applyOverscrollEffect('horizontal', positiveDelta ? 'after' : 'before', overscrollTypeOnce, 1);
+				}
+
+				needToHideScrollbarTrack = true;
 			}
 		}
 
@@ -654,6 +668,10 @@ const useEventWheel = (props, instances) => {
 			}
 
 			scrollContainerHandle.current.scrollToAccumulatedTarget(delta, canScrollVertically, overscrollEffectRequired);
+		}
+
+		if (needToHideScrollbarTrack) {
+			scrollContainerHandle.current.startHidingScrollbarTrack();
 		}
 	}